reorganize a bit
