+++ crates/anodized-core/src/annotate_tests.rs
@@ -1098,7 +1098,7 @@ fn captures_pattern_with_binding_modifier() {
 #[should_panic(expected = "expected an expression")]
 fn captures_missing_assignment_value_is_rejected_by_spec_fields() {
     let input = "captures: Person { name, age } =,";
-    let _: syntax::SpecFields = parse_str(input).unwrap();
+    let _: crate::syntax::spec::SpecFields = parse_str(input).unwrap();
 }
 
 #[test]
@@ -1130,7 +1130,7 @@ fn captures_with_extra_semicolon() {
 
 #[test]
 fn field_value_supports_shorthand_expression() {
-    let spec_fields: syntax::SpecFields = parse_quote! {
+    let spec_fields: crate::syntax::spec::SpecFields = parse_quote! {
         key_1,
         key_2: value,
         key_3: value as expr,
